refactor: convert part of component from .vue to .tsx
