+++ devui/shared/devui-highlight/devui-highlight.tsx
@@ -1,0 +1,12 @@
+import { defineComponent } from 'vue'
+
+export default defineComponent({
+  name: 'd-highlight',
+  props: {
+  },
+  setup(props, ctx) {
+    return () => {
+      return <div>devui-highlight</div>
+    }
+  }
+})
