--- devui/shared/devui-highlight/devui-highlight.vue
@@ -1,13 +1,0 @@
-<template>
-  <div>
-    devui highlight
-  </div>
-</template>
-
-<script>
-import { defineComponent } from 'vue'
-
-export default defineComponent({
-  name: 'd-highlight'
-})
-</script>
